requirements.txt
appdirs==1.4.3
Django==1.10.6
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-sortable==0.3.1
olefile==0.44
packaging==16.8
pdfrw==0.3
Pillow==4.0.0
pkg-resources==0.0.0
psycopg2==2.7
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.0
reportlab==3.4.0
six==1.10.0
django-ckeditor==5.0.3
djangorestframework==3.3.2
factory_boy
